# docs/swagger.yaml
swagger: "2.0"
info:
  description: "これはReseサイトに関するAPIドキュメントです"
  version: "1.0.0"
  title: "Rese API"
...
paths:
...
  /booking:
    post:
      tags:
      - "Booking"
      summary: "店舗予約登録"
      description: "予約登録をします"
      operationId: "Booking"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            store_id:
              type: "string"
              example: "1"
            user_id:
              type: "string"
              example: "1"
            booking_date:
              type: "string"
              example: "2021-05-20"
            booking_time:
              type: "string"
              example: "18:00:00"
            booking_number:
              type: "string"
              example: "2"
      responses:
        "200":
          description: "店舗予約完了"
          schema:
            properties:
              message:
                type: "string"
                example: "Booking successfully"
    put:
      tags:
      - "Booking"
      summary: "予約内容更新"
      description: "予約した内容を更新する"
      operationId: "updateBooking"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            store_id:
              type: "string"
              example: "1"
            user_id:
              type: "string"
              example: "1"
            booking_date:
              type: "string"
              example: "2021-05-25"
            booking_time:
              type: "string"
              example: "20:00:00"
            booking_number:
              type: "string"
              example: "4"
      responses:
        "200":
          description: "予約内容更新完了"
          schema:
            properties:
              message:
                type: "string"
                example: "Booking updated successfully"
    delete:
      tags:
      - "Booking"
      summary: "予約取消"
      description: "予約を取消する"
      operationId: "deleteBooking"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            id:
              type: "string"
              example: "1"
            user_id:
              type: "string"
              example: "1"
      responses:
        "200":
          description: "予約取消完了"
          schema:
            properties:
              message:
                type: "string"
                example: "Favorite deleted successfully"